+++ pcsx2/DEV9/smap.h
@@ -14,4 +14,4 @@ void smap_write32(u32 addr, u32 value);
 
 void smap_readDMA8Mem(u32* pMem, int size);
 void smap_writeDMA8Mem(u32* pMem, int size);
-void smap_async(u32 cycles);
+void smap_async(u64 cycles);
